continuing to factor scenes into renderers + controllers
